final_test_task/requirements.txt
pandas
Faker
sqlalchemy
psycopg2-binary